+++ src/kernels/spgemm/spgemm_common.hpp
@@ -36,7 +36,7 @@ struct Flags {};
 class SpGEMMErrorHandler : public KernelErrorHandler {
   public:
     template <typename IT>
-    static void col_oob(IT col_value, int j, int A_n_cols) {
+    static void col_oob(IT col_value, ULL j, ULL A_n_cols) {
         KernelErrorHandler::col_oob<IT>(col_value, j, A_n_cols, "SpGEMM");
     }
