+++ src/kernels/spmm/spmm_common.hpp
@@ -38,7 +38,7 @@ struct Flags {
 class SpMMErrorHandler : public KernelErrorHandler {
   public:
     template <typename IT>
-    static void col_oob(IT col_value, int j, int A_n_cols) {
+    static void col_oob(IT col_value, ULL j, ULL A_n_cols) {
         KernelErrorHandler::col_oob<IT>(col_value, j, A_n_cols, "SpMM");
     }
